feat: export insightsFromData
