+++ common/reviews/api/chart-advisor.api.md
@@ -7,6 +7,7 @@
 import { ChartID } from '@antv/knowledge';
 import * as DWAnalyzer from '@antv/dw-analyzer';
 import { LevelOfMeasurement } from '@antv/knowledge';
+import { RowData } from '@antv/dw-transform';
 
 // @beta (undocumented)
 export interface Advice {
@@ -86,6 +87,45 @@ export interface FieldInfo extends DWAnalyzer.FieldInfo {
     name: string;
 }
 
+// @beta (undocumented)
+export interface Insight {
+    // (undocumented)
+    fields: string[];
+    // (undocumented)
+    insightProps?: InsightProps;
+    // (undocumented)
+    present?: {
+        data: RowData[];
+        fields: string[];
+    };
+    // (undocumented)
+    type: InsightType | 'SomeInsight';
+}
+
+// @beta (undocumented)
+export const INSIGHT_TYPES: ["Correlation", "Monotonicity", "MajorFactors", "CategoryOutliers", "TimeSeriesOutliers", "OverallTrends", "Seasonality", "ChangePoints"];
+
+// @beta (undocumented)
+export interface InsightProps {
+    // (undocumented)
+    detail?: any;
+    // (undocumented)
+    dimensions?: string[];
+    // (undocumented)
+    measures?: string[];
+    // (undocumented)
+    score?: number;
+}
+
+// @beta (undocumented)
+export function insightsFromData(data: RowData[]): Promise<Insight[]>;
+
+// @beta (undocumented)
+export type InsightType = typeof INSIGHT_TYPES[number];
+
+// @beta (undocumented)
+export const insightWorkers: Partial<Record<InsightType, Worker_2>>;
+
 // @public
 export interface Preferences {
     // (undocumented)
@@ -94,6 +134,11 @@ export interface Preferences {
 
 // @beta
 export function specToLibConfig(advice: Advice, libraryName: ChartLibrary): any;
+
+// @beta (undocumented)
+type Worker_2 = (data: RowData[]) => Insight[] | Promise<Insight[]>;
+
+export { Worker_2 as Worker }
 
 
 ```
